move file and metadaset to dedicated modules
